+++ cloud/acp-spring-cloud-starter/src/main/kotlin/pers/acp/spring/cloud/AcpCloudLogServerClientAutoConfiguration.kt
@@ -1,20 +1,22 @@
 package pers.acp.spring.cloud
 
+import com.fasterxml.jackson.databind.ObjectMapper
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.autoconfigure.AutoConfigureBefore
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression
-import org.springframework.cloud.stream.annotation.EnableBinding
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.cloud.stream.config.BindingProperties
 import org.springframework.cloud.stream.config.BindingServiceConfiguration
 import org.springframework.cloud.stream.config.BindingServiceProperties
+import org.springframework.cloud.stream.function.StreamBridge
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.util.MimeTypeUtils
 import pers.acp.spring.cloud.conf.AcpCloudLogServerClientConfiguration
-import pers.acp.spring.cloud.conf.AcoCloudLogServerConfiguration
-import pers.acp.spring.cloud.log.producer.LogOutput
-import pers.acp.spring.cloud.log.producer.LogProducer
+import pers.acp.spring.cloud.conf.AcpCloudLogServerConfiguration
 import pers.acp.spring.cloud.log.LogConstant
+import pers.acp.spring.cloud.log.producer.LogBridge
+import pers.acp.spring.cloud.log.producer.StreamLogBridge
 
 import javax.annotation.PostConstruct
 
@@ -27,24 +29,24 @@ import javax.annotation.PostConstruct
 @Configuration(proxyBeanMethods = false)
 @ConditionalOnExpression("'\${acp.cloud.log-server.client.enabled}'.equals('true')")
 @AutoConfigureBefore(BindingServiceConfiguration::class, AcpCloudLogAutoConfiguration::class)
-@EnableBinding(LogOutput::class)
 class AcpCloudLogServerClientAutoConfiguration @Autowired
-constructor(private val acoCloudLogServerConfiguration: AcoCloudLogServerConfiguration,
-            private val acpCloudLogServerClientConfiguration: AcpCloudLogServerClientConfiguration,
-            private val bindings: BindingServiceProperties) {
-
+constructor(
+    private val acpCloudLogServerConfiguration: AcpCloudLogServerConfiguration,
+    private val acpCloudLogServerClientConfiguration: AcpCloudLogServerClientConfiguration,
+    private val bindingServiceProperties: BindingServiceProperties
+) {
     /**
      * 初始化日志消息生产者
      */
     @PostConstruct
     fun init() {
         if (acpCloudLogServerClientConfiguration.enabled) {
-            if (this.bindings.bindings[LogConstant.OUTPUT] == null) {
-                this.bindings.bindings[LogConstant.OUTPUT] = BindingProperties()
+            if (this.bindingServiceProperties.bindings[LogConstant.OUTPUT] == null) {
+                this.bindingServiceProperties.bindings[LogConstant.OUTPUT] = BindingProperties()
             }
-            this.bindings.bindings[LogConstant.OUTPUT]?.let {
+            this.bindingServiceProperties.bindings[LogConstant.OUTPUT]?.let {
                 if (it.destination == null || it.destination == LogConstant.OUTPUT) {
-                    it.destination = acoCloudLogServerConfiguration.destination
+                    it.destination = acpCloudLogServerConfiguration.destination
                 }
                 it.contentType = MimeTypeUtils.APPLICATION_JSON_VALUE
             }
@@ -52,6 +54,7 @@ constructor(private val acoCloudLogServerConfiguration: AcoCloudLogServerConfigu
     }
 
     @Bean
-    fun logProducer(logOutput: LogOutput): LogProducer = LogProducer(logOutput)
-
+    @ConditionalOnMissingBean(LogBridge::class)
+    fun streamLogBridge(streamBridge: StreamBridge, objectMapper: ObjectMapper): LogBridge =
+        StreamLogBridge(streamBridge, objectMapper, LogConstant.OUTPUT)
 }