+++ cloud/acp-spring-cloud-starter/src/main/kotlin/pers/acp/spring/cloud/log/producer/LogProducer.kt
@@ -1,7 +1,0 @@
-package pers.acp.spring.cloud.log.producer
-
-/**
- * @author zhang by 14/01/2019 17:08
- * @since JDK 11
- */
-class LogProducer(val logOutput: LogOutput)